14-openagents/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openagents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "geopy>=2.4.1",
    "google-auth>=2.23.0",
    "googlemaps>=4.10.0",
    "openai>=2.8.0",
    "openai-agents>=0.5.1",
    "pytest>=9.0.1",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.1",
    "streamlit>=1.28.0",
    "yfinance>=0.2.66",
]

[tool.setuptools]
packages = ["asdrp"]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.12.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
